# src/nemo/apigateway-service/swagger/openapi-consolidated.yaml
openapi: 3.0.3
info:
  title: AgentStudio Services API
  version: 1.0.0
  description: Consolidated API documentation for all AgentStudio services (Config
    Service)
servers:
- url: http://localhost:8080
  description: AgentStudio Gateway (local)
- url: https://agentstudio.local:8443
  description: AgentStudio Gateway (local development)
security:
- bearerAuth: []
- oauth2:
  - openid
  - profile
  - email
paths:
...
  /config/api/v1/deployments/{deploymentId}:
    get:
      summary: Get deployment
      tags:
      - Deployments
      description: Gets a deployment by ID
      operationId: getDeployment
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update deployment
      tags:
      - Deployments
      description: Updates a deployment
      operationId: updateDeployment
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeploymentRequest'
      responses:
        '200':
          description: Updated deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Deregister deployment
      tags:
      - Deployments
      description: Deregisters a deployment
      operationId: deleteDeployment
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deployment deregistered
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /config/api/v1/deployments/{deploymentId}/config:
    get:
      summary: Get deployment config
      tags:
      - Deployments
      description: Gets full deployment configuration
      operationId: getDeploymentConfig
      parameters:
      - name: deploymentId
        in: path
        required: true
        schema:
          type: string
      - name: since
        in: query
        schema:
          type: integer
        description: Config version for incremental updates
      responses:
        '200':
          description: Deployment configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentConfigResponse'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /config/api/v1/buckets/{projectId}/{bucketName}/routing:
    get:
      summary: Get bucket routing
      tags:
      - Deployments
      description: Gets routing information for a bucket (all deployments serving
        it)
      operationId: getBucketRouting
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: bucketName
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Routing information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BucketRoutingResponse'
        '404':
          description: Bucket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
...
    Deployment:
      type: object
      properties:
        id:
          type: string
        region:
          type: string
        endpoint:
          type: string
        capacity:
          type: object
          properties:
            max_buckets:
              type: integer
            max_storage_tb:
              type: integer
        storage_classes:
          type: array
          items:
            type: string
          description: List of StorageClass names supported by this deployment
        capabilities:
          type: array
          items:
            type: string
        registered_at:
          type: string
          format: date-time
        last_health_check:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - healthy
          - unhealthy
          - unknown
...
    UpdateDeploymentRequest:
      type: object
      properties:
        region:
          type: string
        endpoint:
          type: string
        capacity:
          type: object
          properties:
            max_buckets:
              type: integer
            max_storage_tb:
              type: integer
        capabilities:
          type: array
          items:
            type: string
        storage_classes:
          type: array
          items:
            type: string
          description: List of StorageClass names supported by this deployment
...
    DeploymentConfigResponse:
      type: object
      properties:
        deployment_id:
          type: string
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/BucketConfig'
        config_version:
          type: integer
    BucketConfig:
      type: object
      properties:
        namespace_id:
          type: string
        bucket_name:
          type: string
        region:
          type: string
        volume_info:
          type: object
          properties:
            type:
              type: string
            endpoint:
              type: string
            mount_options:
              type: array
              items:
                type: string
            provisioning_mode:
              type: string
              enum:
              - static
              - dynamic
            storage_class_name:
              type: string
            storage_size:
              type: string
            parameters:
              type: object
              additionalProperties:
                type: string
            access_modes:
              type: array
              items:
                type: string
                enum:
                - ReadWriteOnce
                - ReadWriteMany
                - ReadOnlyMany
                - ReadWriteOncePod
        auth_info:
          type: object
          properties:
            type:
              type: string
            username:
              type: string
            password_encrypted:
              type: string
          additionalProperties: true
        protocol:
          type: string
        role:
          type: string
          enum:
          - primary
          - secondary
        other_deployments:
          type: array
          items:
            type: object
            properties:
              deployment_id:
                type: string
              role:
                type: string
                enum:
                - primary
                - secondary
    BucketRoutingResponse:
      type: object
      properties:
        bucket_name:
          type: string
        namespace_id:
          type: string
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/RoutingDeploymentInfo'
        routing_strategy:
          type: string
          enum:
          - load_balance
          - failover
          - geographic
        updated_at:
          type: string
          format: date-time
    RoutingDeploymentInfo:
      type: object
      properties:
        deployment_id:
          type: string
        role:
          type: string
          enum:
          - primary
          - secondary
        priority:
          type: integer
        endpoint:
          type: string
        health_status:
          type: string
          enum:
          - healthy
          - unhealthy
          - unknown
        load_balance_weight:
          type: integer
        last_health_check:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token obtained from Keycloak. You can get a token
        by logging into the GUI and copying the token from browser developer tools,
        or by using the OAuth2 flow below.
    oauth2:
      type: oauth2
      description: OAuth2 authentication via Keycloak. Click Authorize and login with
        your credentials.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.agentstudio.local:8443/realms/nemo/protocol/openid-connect/auth
          tokenUrl: https://auth.agentstudio.local:8443/realms/nemo/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email address